Streamlined API with Thor

+++ api/API.rb
@@ -1,47 +1,232 @@
+#!/usr/bin/env ruby
+
 require 'json'
 require 'net/http'
 require 'uri'
 require 'csv'
+require 'thor'
 
-## Script to call NetBox API
+## Script to interact with the NetBox API
 #
-# Arguments passed to this script will call class methods
+# Uses Thor to manage input
 # 
 # Example:
 # 
 # ruby API.rb dns_combine
 #
-# Make sure to set this to wherever the server is running:
-base = "http://localhost"
-fun = ARGV[0]
 
 ## Class to call the Netbox API and get various results
+# Token is passed in an Environment Variable
+# export token=.......
+# The netbox URL can optionally be passed in an environment variable as well
+# export netbox_url=http://...
 # Takes an optional "options" hash which is passed to the API request
 # The default options hash returns all results without filtering
-class NetboxAPI
-	@@options = {'limit' => 0, 'offset' => 0}
-
-	def initialize(base_url, options = @@options)
-		@base_url = base_url
-		@@options = options
+class NetboxAPI < Thor
+	@@token = ENV['token']
+	if ENV['netbox_url'].nil? 
+		@@base_url = 'localhost'
+	else
+		@@base_url = ENV['netbox_url']
 	end
+	@@http_options = {'limit' => 0, 'offset' => 0}
 
-	# Call the API
-	# Return: JSON
-	def call(path, options = nil)
-		options = @@options if options.nil?
-		unless options.instance_of?(Hash)
+	# Send a GET request
+	# Return: Array
+	desc "get path ...options, base_url", "Send a GET Request"
+	method_option :getspeak, :desc => "Print the output here"
+	def get(path, http_options = nil, base_url = nil)
+		speak = options[:getspeak]
+		http_options = @@http_options if http_options.nil?
+		unless http_options.instance_of?(Hash)
 			return nil
 		end
-		options = URI.encode_www_form(options)
-		url = "#{@base_url}/api/#{path}/?#{options}"
+		base_url = @@base_url if base_url.nil?
+		http_options = URI.encode_www_form(http_options)
+		url = "http://#{base_url}/api/#{path}/?#{http_options}"
 		uri = URI.parse(url)
 		response = Net::HTTP.get_response(uri)
+		results = JSON.parse(response.body)['results']
 		if response.code == '200'
-		  JSON.parse(response.body)['results']
+			puts results if speak
+			results
 		else
 		  puts "Invalid response #{response.code}"
 		end
+	end
+
+	# Send a POST request
+	# Return: HTTP Response Body
+	desc "post path post_body", "Send a POST Request"
+	method_option :postspeak, :desc => "Print the output here"
+	def post(path, post_body, base_url = nil)
+		speak = options[:postspeak]
+		base_url = @@base_url if base_url.nil?
+		url = "http://#{base_url}/api/#{path}/"
+		uri = URI.parse(url)
+		http = Net::HTTP.new(uri.host, uri.port)
+		req = Net::HTTP::Post.new(uri.request_uri)
+		req['Content-Type'] = 'application/json'
+		req['Accept'] = 'application/json'
+		req['Authorization'] = "Token #{@@token}"
+		req.body = post_body.to_json
+		results = http.request(req).body
+		if speak
+			puts results
+		else
+			results
+		end
+	end
+
+	# Send a PATCH request
+	# Return: HTTP Response Body
+	desc "patch ...path", "Send a PATCH Request"
+	method_option :patchspeak, :desc => "Print the output here"
+	def patch(path, id, post_body)
+		speak = options[:patchspeak]
+		url = "http://#{@@base_url}/api/#{path}/#{id}/"
+		uri = URI.parse(url)
+		http = Net::HTTP.new(uri.host, uri.port)
+		req = Net::HTTP::Patch.new(uri.request_uri)
+		req['Content-Type'] = 'application/json'
+		req['Accept'] = 'application/json'
+		req['Authorization'] = "Token #{@@token}"
+		req.body = post_body.to_json
+		results = http.request(req).body
+		if speak
+			puts results
+		else
+			results
+		end
+	end
+
+	# Add a virtual machine
+	# 
+	# Assumes Active status
+	# Assumes Centos platform
+	# 
+	# Returns: HTTP Status Code
+	desc "add_virtual_machine name clusterid", "Add a Virtual Machine with name and clusterid"
+	def add_virtual_machine(name, clusterid)
+		path = 'virtualization/virtual-machines'
+		post_body = {
+								name: name,
+								status: 1,
+								cluster: clusterid,
+								platform: 2
+							}
+		post(path, post_body)
+	end
+
+	# Bulk Add Virtual Machines
+	# 
+	# 
+	# Returns: HTTP Status Code
+	desc "bulk_add_virtual_machines filename", "Bulk add VM's from a File"
+	def bulk_add_virtual_machines(filename)
+		listof = File.readlines(filename)
+		listof.each do |vm|
+			array = vm.split(',')
+			name = array[0]
+			clusterid = array[1]
+			# Get the vmid of the vm with name
+			#vm_id = get_vm_id_from_name(name)
+			# Create the IP address - pass as an array
+			#response = add_ip_address(ip)
+			#ip_id = JSON.parse(response)['id']
+			# Set the IP to the Virtual Machine
+			add_virtual_machine(name, clusterid)
+		end
+	end
+
+	# Update a virtual machine's Primary IP address
+	# 
+	# Returns: HTTP Status Body
+	desc "update_virtual_machine_ip vm_id ip_id", "Update Virtual Machine IP"
+	def update_virtual_machine_ip(vm_id, ip_id)
+		path = 'virtualization/virtual-machines'
+		post_body = {
+								primary_ip4: ip_id,
+							}
+		patch(path, vm_id, post_body)
+	end
+
+	# Add a virtual machine interface for each virtual machine
+	# 
+	# Use case:
+	# There's not one by default and all of my VM's are on eth0
+	# 
+	# Assumes Virtual Interface
+	# Assumes Interface is Enabled
+	# 
+	# Returns: HTTP Status Code
+	desc "bulk_add_virtual_machine_interface interface", "Add Virtual Machine Interfaces in Bulk"
+	def bulk_add_virtual_machine_interface(interface)
+		path = 'virtualization/virtual-machines'
+		vms = get(path)
+		vms.each do |vm|
+			vm_id = vm['id']
+			post_body = {
+						  "name": interface,
+						  "virtual_machine": vm_id,
+						  "form_factor": 0,
+						  "enabled": true
+							}
+			path = 'virtualization/interfaces'
+			post(path, post_body)
+		end
+	end
+
+	# Add an IP to a list of Virtual Machines
+	# 
+	# Use case:
+	# I have a long list of VM's and I don't want to set their
+	# IP's manually
+	# 
+	# File is comma-seperated hostname,ip
+	# 
+	# Returns: HTTP Status Code
+	desc "bulk_set_virtual_machine_ip filename interface", "Set Virtual Machine IP's in Bulk"
+	def bulk_set_virtual_machine_ip(filename, interface)
+		listof = File.readlines(filename)
+		listof.each do |vm|
+			array = vm.split(',')
+			name = array[0]
+			ip = array[1]
+			# Get the vmid of the vm with name
+			vm_id = get_vm_id_from_name(name)
+			# Create the IP address - pass as an array
+			response = add_ip_address(ip)
+			ip_id = JSON.parse(response)['id']
+			# Set the IP to the Virtual Machine
+			update_virtual_machine_ip(vm_id, ip_id)
+		end
+	end
+
+	# Get a Virtual Machine ID from it's name
+	# 
+	# Returns: Integer
+	desc "get_vm_id_from_name name", "Set Virtual Machine IP's in Bulk"
+	def get_vm_id_from_name(name)
+		path = 'virtualization/virtual-machines'
+		http_options = { 'name': name }
+		vm = get(path, http_options)
+		vm[0]['id']
+	end
+
+	# Add an IP address
+	# 
+	# Assumes Active
+	#
+	# Returns: HTTP Status Code
+	desc "add_ip_address ip", "Set Virtual Machine IP's"
+	def add_ip_address(ip)
+		path = 'ipam/ip-addresses'
+		post_body = {
+						"address": ip,
+						"status": 1
+		}
+		post(path, post_body)
 	end
 
 	## DNS methods
@@ -56,38 +241,49 @@ class NetboxAPI
 	# Optional limit argumument to limit the number of results given
 	# 
 	# Returns: Array
+	desc "dns_vms limit", "Get DNS Information for VM's"
+	method_option :dnsvmsspeak, :aliases => '-s', :desc => "Print the output here"
 	def dns_vms(limit = 0)
+		speak = options[:dnsvmsspeak]
 		array = Array.new
 		path = 'virtualization/virtual-machines'
 		options = {'status' => 1, 'limit' => limit, 'offset' => 0}
-		vms = call(path,options)
+		vms = get(path, options)
 		vms.each do |vm|
 			hash = Hash.new
-			id = vm['id']
+			vm_id = vm['id']
 			hash['name'] = vm['name']
 			# Get IP address by getting the IP ID
 			ip_id = vm['primary_ip4']
 			unless ip_id.nil?
 				ip_options = {'id__in' => ip_id}
-				ipam = call("ipam/ip-addresses",ip_options)
+				ipam = get("ipam/ip-addresses",ip_options)
 				hash['ip'] = ipam[0]['address'].split(/\//).first
 			end
-			# Cycle thru all services to find mine until this is fixed
-			services = call("ipam/services")
-			unless services.empty?
+			hash['type'] = 'A'
+			# Get the services associated with this VM
+			options = {'virtual_machine_id' => vm_id}
+			services = get("ipam/services", options)
+			unless services.nil?
 				services.each do |service|
-					if service['virtual_machine']
-						if id == service['virtual_machine']['id']
-							ip = service['ipaddresses'][0]['address'].split(/\//).first
-							hash['servicename'] = service['name']
-							hash['serviceip'] = ip
-						end
+					if service['ipaddresses'].empty?
+						ip = service['description']
+						hash['servicetype'] = 'CNAME'
+					else
+						ip = service['ipaddresses'][0]['address'].split(/\//).first
+						hash['servicetype'] = 'A'
 					end
+					hash['servicename'] = service['name']
+					hash['serviceip'] = ip
 				end
 			end
 			array << hash
 		end
-		array
+		if speak
+			puts array.inspect
+		else
+			array
+		end
 	end
 
 	# Returns node names, ips, and related service names and ips for use in order
@@ -96,27 +292,35 @@ class NetboxAPI
 	# Optional limit argumument to limit the number of results given
 	# 
 	# Returns: Array
-	def dns_physical(limit = 0)
+	desc "dns_physical ...limit", "Get DNS Information for Physical Nodes"
+	def dns_physical(limit = 1)
 		array = Array.new
 		path = 'dcim/devices'
 		options = {'status' => 1, 'limit' => limit, 'offset' => 0}
-		devices = call(path, options)
+		devices = get(path, options)
 		devices.each do |device|
 			unless device['primary_ip'].nil?
 				hash = Hash.new
 				hash['name'] = device['name']
 				hash['ip'] = device['primary_ip']['address'].split(/\//).first
+				hash['type'] = 'A'
 				service_options = {'device_id' => device['id']}
-				services = call("ipam/services",service_options)
-				unless services.empty?
+				services = get("ipam/services",service_options)
+				unless services.nil?
 					services.each do |service|
-						ip = service['ipaddresses'][0]['address']
+						if service['ipaddresses'].empty?
+							ip = service['description']
+							hash['servicetype'] = 'CNAME'
+						else
+							ip = service['ipaddresses'][0]['address'].split(/\//).first
+							hash['servicetype'] = 'A'
+						end
 						hash['servicename'] = service['name']
-						hash['serviceip'] = ip.split(/\//).first
+						hash['serviceip'] = ip
 					end
 				end
-				array << hash
 			end
+			array << hash
 		end
 		array
 	end
@@ -124,24 +328,25 @@ class NetboxAPI
 	# Combine Virtual Machines and Hardware Device DNS Information into a single array
 	# 
 	# Returns: True, Writes to CSV
-	def dns_combine
+	desc "dns_combine limit", "Export DNS Information for Hosts"
+	def dns_combine(limit = 0)
 		filename='data.csv'
 		# Clean up the old file jic
-		File.delete(filename)
-		vms = dns_vms
-		physical = dns_physical
+		File.delete(filename) if File.exist?(filename)
+		vms = dns_vms(limit)
+		physical = dns_physical(limit)
 		array = [*vms,*physical]
 		array.each do |d|
-			open(filename, 'a+') { |f| f << "#{d['name']},#{d['ip']}\n" }
+			open(filename, 'a+') { |f| f << "#{d['name']},#{d['ip']},#{d['type']}\n" }
 			if d['servicename']
-				open(filename, 'a+') { |f| f << "#{d['servicename']},#{d['serviceip']}\n" }
+				open(filename, 'a+') { |f| f << "#{d['servicename']},#{d['serviceip']},#{d['servicetype']}\n" }
 			end
 		end
 		# De-duplicate services
 		dups = File.readlines(filename)
 		unqd = dups.uniq
 		unqd = unqd.sort
-		File.open(filename, 'w') { |f| f << "name,ip\n" }
+		File.open(filename, 'w') { |f| f << "name,ip,type\n" }
 		File.open(filename, "a+") do |f|
   		f.puts(unqd)
 		end
@@ -149,9 +354,9 @@ class NetboxAPI
 	end
 end
 
-API = NetboxAPI.new(base)
-if API.respond_to? fun
-	puts API.public_send(fun)
-else
-	puts "No such method"
+if ENV['token'].nil?
+	puts "Token must be set"
+	puts "export token=......"
+	exit 1
 end
+NetboxAPI.start(ARGV)
